requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==1.12.3
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
ciso8601==2.2.0
colorlog==6.6.0
distlib==0.3.4
filelock==3.6.0
frozenlist==1.3.0
hikari==2.0.0.dev108
hikari-lightbulb==2.2.1
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
platformdirs==2.5.1
py==1.11.0
pycares==4.1.2
pycparser==2.21
pyparsing==3.0.7
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.8.0
urllib3==1.26.9
uvloop==0.16.0
virtualenv==20.14.0
yarl==1.7.2
